chore(script): Add script to display subdomain information
Added a Bash script to display information about subdomains. The script takes a domain name and an optional subdomain as arguments. It uses dig and awk commands to retrieve and parse DNS information for the specified subdomains. If no subdomain is provided, information for default subdomains (www, lb-01, web-01, web-02) is displayed.