feat(API): Set up Nginx and Gunicorn for AirBnB clone v3
This commit includes the setup of Nginx to route /api/ to a Gunicorn instance listening on port 5002. The setup allows Nginx to serve the page both locally and on its public IP on port 80. Gunicorn is bound to api/v1/app.py for testing the setup. The Nginx config file has been uploaded as 4-app_server-nginx_config.